moved out branch ordering functions to tracing utils
